% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_ngrok.R
\name{auth_ngrok}
\alias{auth_ngrok}
\title{authorize ngrok using your personal token https://dashboard.ngrok.com/get-started/your-authtoken}
\usage{
auth_ngrok(token)
}
\arguments{
\item{token}{a string token from https://dashboard.ngrok.com/get-started/your-authtoken}
}
\description{
authorize your online ngrok subscription to work on your computer
}
