% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livestream_start.R
\name{livestream_start}
\alias{livestream_start}
\title{start the livestream}
\usage{
livestream_start(
  hostname = NULL,
  password = NULL,
  user = NULL,
  port = 4040,
  scheme = "http"
)
}
\arguments{
\item{hostname}{custom domain}

\item{password}{password for the group}

\item{user}{login for the group}

\item{port}{port for tunnel}

\item{scheme}{use http or https (http seems to be the only thing that works at the moment)}
}
\description{
authorize your online ngrok subscription to work on your computer
}
\details{
for some reason, https ends up showing a blank page, so forcing http scheme is the only way to get the custom url to work right
}
