% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livestream_stop.R
\name{livestream_stop}
\alias{livestream_stop}
\title{stop the livestream}
\usage{
livestream_stop(id = NULL)
}
\arguments{
\item{id}{takes an id string from livestream_start return}
}
\description{
authorize your online ngrok subscription to work on your computer
}
